+++ usecases/fetchMeteors.ts
@@ -1,38 +1,5 @@
 import { getMeteorData } from '../repository/index.ts'
-
-interface FetchMeteorsResponse {
-  meteorData: Array<SimpleMeteor>
-  wereDangerousMeteors?: boolean
-  count?: number
-}
-
-interface EstimatedDiameter {
-  meters: number
-}
-
-interface RelativeVelocity {
-  kilometers_per_second: string
-}
-
-interface CloseApproachData {
-  close_approach_date_full: string
-  relative_velocity: RelativeVelocity
-}
-
-interface Meteor {
-  id: string
-  name: string
-  estimated_diameter: EstimatedDiameter
-  is_potentially_hazardous_asteroid: boolean
-  close_approach_data: Array<CloseApproachData>
-}
-
-interface SimpleMeteor
-  extends Pick<Meteor, 'id' | 'name' | 'is_potentially_hazardous_asteroid'> {
-  diameter: number
-  close_approach_date_full: string
-  relative_velocity: number
-}
+import { FetchMeteorsResponse, Meteor, SimpleMeteor } from '../types/meteor.ts'
 
 export const fetchMeteorsUseCase = async (
   startDate: Date,
